+++ lib/jinja.dart
@@ -1,5 +1,3 @@
-library jinja;
-
 export 'package:jinja/src/environment.dart';
 export 'package:jinja/src/exceptions.dart';
 export 'package:jinja/src/loaders.dart';