+++ lib/loaders.dart
@@ -1,5 +1,3 @@
-library jinja.loaders;
-
 import 'dart:convert';
 import 'dart:io';
 
@@ -72,7 +70,7 @@ class FileSystemLoader extends Loader {
     var file = findFile(path);
 
     if (file == null) {
-      throw TemplateNotFound(name: path);
+      throw TemplateNotFound(path: path);
     }
 
     return file.readAsStringSync(encoding: encoding);